+++ tests/JsonLoaderTest.php
@@ -8,8 +8,7 @@ declare(strict_types=1);
 
 namespace SugiPHP\Config\Tests;
 
-use SugiPHP\Config\JsonLoader as Loader;
-use SugiPHP\Config\FileLocator as Locator;
+use SugiPHP\Config\Loader\JsonLoader as Loader;
 use PHPUnit\Framework\TestCase;
 
 class JsonLoaderTest extends TestCase
@@ -17,7 +16,7 @@ class JsonLoaderTest extends TestCase
     public function testJsonLoaderIsLoaderInterface()
     {
         $loader = new Loader();
-        $this->assertInstanceOf("\SugiPHP\Config\LoaderInterface", $loader);
+        $this->assertInstanceOf("\SugiPHP\Config\Loader\LoaderInterface", $loader);
     }
 
     public function testReturnsNull()
@@ -39,8 +38,7 @@ class JsonLoaderTest extends TestCase
     {
         $testArr = include __DIR__."/config/test.php";
 
-        $locator = new Locator(array(__DIR__, __DIR__."/config"));
-        $loader = new Loader($locator);
+        $loader = new Loader(array(__DIR__, __DIR__."/config"));
         $this->assertNull($loader->load("nosuchfile"));
         $this->assertEquals($testArr, $loader->load("config/test.json"));
         $this->assertEquals($testArr, $loader->load("config/test"));