+++ tests/config2/site.php
@@ -1,0 +1,3 @@
+<?php
+
+return array("name" => "site in config2");